Refactor, add triangular lattice subclass
